requirements.txt
gradio>=3.50.2
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
urllib3>=2.0.7
regex>=2023.10.3
jieba>=0.42.1
